first commit with fetch

+++ src/js/component/home.jsx
@@ -1,8 +1,7 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 
-//include images into your bundle
-import rigoImage from "../../img/rigo-baby.jpg";
-
+const urlBase = "http://assets.breatheco.de/apis/fake/todos/user/";
+const apiUsername = "nikodemo69";
 //create your first component
 
 let initialTasks = [];
@@ -11,6 +10,41 @@ const Home = () => {
   const [tasks, setTasks] = useState(initialTasks);
   const [taskName, setTaskName] = useState("");
 
+
+// Buscamos la lista de tareas. METODO GET para obtener la info y traer la lista de tareas.
+const fetchTodoApi = async() => {
+  try{
+    const response = await fetch(`${urlBase}${apiUsername}`);
+    console.log(response);
+    const data = await response.json(); // Traducimos la respuesta a json
+    console.log(data);
+    setTasks(data)
+  }catch(error){ 
+    console.log(error);
+  }
+};
+
+// Usamos esta funcion para actualizar la informacion de la lista de tareas
+const updateToDoList = async(tasks) => {
+  console.log(tasks)
+  try{
+    const response = await fetch(`${urlBase}${apiUsername}`, {
+      method:"PUT",
+      body:JSON.stringify(tasks),
+      headers: {
+        "Content-Type": "application/json"
+      }
+    });
+    fetchTodoApi();
+  }catch(error){
+    console.log(error)
+  }
+};
+
+
+
+  
+  //Lleva registro de el Todo
   const handleTaskChange = (event) => {
     setTaskName(event.target.value);
   };
@@ -21,10 +55,11 @@ const Home = () => {
 
     if (event.key === "Enter") {
       const newTask = {
-        name: taskName,
+        label: taskName, 
+        done: false
       };
       const newTasks = [...tasks, newTask];
-      setTasks(newTasks);
+      updateToDoList(newTasks);
       setTaskName("");
     }
   };
@@ -33,8 +68,17 @@ const Home = () => {
   const deleteItem = (index) => {
     const newInitialTasks = tasks.filter((task, i) => index !== i);
     console.log(newInitialTasks);
-    setTasks(newInitialTasks);
+    updateToDoList(newInitialTasks);
   };
+
+console.log("Antes del useEffect")
+  useEffect(() => {
+    console.log("Dentro del useEffect")
+    fetchTodoApi();
+  }, [])
+console.log("Despues del useEffect")
+
+
 
   return (
     <div className="d-flex justify-content-center align-items-center">
@@ -44,7 +88,7 @@ const Home = () => {
           <div className="card-body w-100 backgroundTest container-fluid">
             <ul className="list-group list-group-flush w-100">
               <input
-                class="list-group-item opacity-75 fs-4"
+                className="list-group-item opacity-75 fs-4"
                 type="text"
                 placeholder="What needs to be done?"
                 onChange={(event) => handleTaskChange(event)}
@@ -52,27 +96,34 @@ const Home = () => {
                 value={taskName}
               ></input>
               {tasks.map((task, index) => {
-                return (
-                  <li
-                    id="wholeTask"
-                    className="list-group-item d-flex justify-content-between p-2 fs-4 opacity-75"
-                    key={`${task.name}-${index}`}
-                  >
-                    <div id="taskName" className="me-auto">
-                      {task.name}
-                    </div>
-                    <div id="XBotton" className="opacity-75">
-                      <i
-                        class="fas fa-times"
-                        onClick={() => deleteItem(index)}
-                      ></i>
-                    </div>
-                  </li>
-                );
+                if (task.done != true) {
+                  return (
+                    <li
+                      id="wholeTask"
+                      className="list-group-item d-flex justify-content-between p-2 fs-4 opacity-75"
+                      key={`${task.label}-${index}`}
+                    >
+                      <div id="taskName" className="me-auto">
+                        {task.label}
+                      </div>
+                      <div id="XBotton" className="opacity-75">
+                        <i
+                          className="fas fa-times"
+                          onClick={() => deleteItem(index)}
+                        ></i>
+                      </div>
+                    </li>
+                  );
+                } 
+
               })}
             </ul>
           </div>
-          <div className="card-footer">{tasks.length} items left</div>
+          <div className="card-footer">{tasks.length - 1} items left
+          </div>
+          <button onClick={() => updateToDoList([{label:"No mostrar", done: true}])} type="button" class="btn btn-warning">Clear list</button>
+         
+
         </div>
       </div>
     </div>
@@ -84,3 +135,5 @@ export default Home;
 {
   /* <button onClick={() => deleteItem(index)}>X</button> */
 }
+
+//linea 81, cambie task.name por task.label y no me funciona ahora.
